refactor(catalog-cli): carve collections command family into catalog_cmds (nexus-whh61.4)
Second continuation family. Carve the collection-management commands
(backfill-collections, collection-name, rename-collection, collection-gc)
out of commands/catalog.py into commands/catalog_cmds/collections.py,
mirroring the owners/backfill/links/backups pattern: plain @click.command +
register(group); commands.catalog calls _collections_cmds.register(catalog).
'nx catalog backfill-collections' (etc.) resolve identically — names,
options, output unchanged; 47 total catalog commands preserved.

Behaviour-preserving: bodies moved verbatim (incl. lazy make_t3 /
_BYPASS_SCHEMA_PREFIXES / corpus / commands.collection imports and the
rename data-plane + projection-recovery WARN path). _get_catalog /
_get_catalog_writer reached via the nexus.commands.catalog module object
inside each body (acyclic + monkeypatch-safe). collections.py carries its
own structlog logger for backfill-collections' _log.warning. No test imports
these by symbol; _get_owner_for (used only by later doctor commands) left in
catalog.py.

Three contract pins in test_catalog_cli.py::TestWhh61CollectionsCarve:
registration, callbacks-in-carved-module, behavioural module-routed
_get_catalog via backfill-collections. Guard suites green (128: cli +
collection-gc + backfill-collections + rename-collection).
